Mint Magazine: Web Fundamentals I Final
For my final project, I designed a basic minimalistic magazine layout for an idea I've had for awhile. As a journalist, I wanted to combine my creative intents with the basic coding we've learned this semester. This is a 'sketch' of what Mint magazine could potentially be in the next few years. Mint contains html, css, and javascript features including images and a film report from the MoMa written by myself.